+++ src/main/java/com/insane96mcp/shieldsplus/setup/Client.java
@@ -20,23 +20,10 @@ public class Client {
 
     @SubscribeEvent
     public static void onStitch(TextureStitchEvent.Pre event) {
-        //TODO Check if I can loop the SPItems.SHIELDS to read all the textures from there=
         for (RegistryObject<SPShieldItem> shieldItem : SPItems.SHIELDS) {
             event.addSprite(shieldItem.get().material.material.texture());
             event.addSprite(shieldItem.get().material.noPatternMaterial.texture());
         }
-        /*if (event.getAtlas().location().equals(InventoryMenu.BLOCK_ATLAS)) {
-            event.addSprite(SPShieldMaterials.WOODEN.material.texture());
-            event.addSprite(SPShieldMaterials.WOODEN.noPatternMaterial.texture());
-            event.addSprite(SPShieldMaterials.STONE.material.texture());
-            event.addSprite(SPShieldMaterials.STONE.noPatternMaterial.texture());
-            event.addSprite(SPShieldMaterials.GOLDEN.material.texture());
-            event.addSprite(SPShieldMaterials.GOLDEN.noPatternMaterial.texture());
-            event.addSprite(SPShieldMaterials.DIAMOND.material.texture());
-            event.addSprite(SPShieldMaterials.DIAMOND.noPatternMaterial.texture());
-            event.addSprite(SPShieldMaterials.NETHERITE.material.texture());
-            event.addSprite(SPShieldMaterials.NETHERITE.noPatternMaterial.texture());
-        }*/
     }
 
     private static void initShields() {